[fix] fixed gws delete_all user

+++ app/controllers/concerns/gws/crud_filter.rb
@@ -128,9 +128,11 @@ module Gws::CrudFilter
     entries.each do |item|
       if item.allowed?(:delete, @cur_user, site: @cur_site)
         item.attributes = fix_params
-        if item.deletion_unlocked? && item.disabled?
-          item.destroy
-          next
+        if item.is_a?(Gws::User)
+          if item.deletion_unlocked? && item.disabled?
+            item.destroy
+            next
+          end
         end
         next if item.disable
       else
